# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LendyPH API",
        "description": "REST API for LendyPH Lending Application",
        "contact": {
            "name": "LendyPH",
            "email": "support@lendyph.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/borrowers": {
            "get": {
                "tags": [
                    "Borrowers"
                ],
                "summary": "List borrowers",
                "description": "Get a paginated list of borrowers with search and filters",
                "operationId": "e65b6b86eff53b9b95c45934b5a6af08",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        }
                    },
                    {
                        "name": "branch_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated borrower list"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Borrowers"
                ],
                "summary": "Create borrower",
                "description": "Create a new borrower profile",
                "operationId": "01f5494e407480a6986b39fb6e66b27e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "branch_id"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "middle_name": {
                                        "type": "string",
                                        "example": "Santos"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Dela Cruz"
                                    },
                                    "suffix": {
                                        "type": "string",
                                        "example": "Jr."
                                    },
                                    "birthdate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-15"
                                    },
                                    "civil_status": {
                                        "type": "string",
                                        "enum": [
                                            "single",
                                            "married",
                                            "widowed",
                                            "separated",
                                            "divorced"
                                        ]
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ]
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "contact_number": {
                                        "type": "string",
                                        "example": "09171234567"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@email.com"
                                    },
                                    "employer_or_business": {
                                        "type": "string"
                                    },
                                    "monthly_income": {
                                        "type": "number",
                                        "example": 25000
                                    },
                                    "branch_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Borrower created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/borrowers/{id}": {
            "get": {
                "tags": [
                    "Borrowers"
                ],
                "summary": "Show borrower",
                "description": "Get full borrower profile with co-makers and documents",
                "operationId": "d8565712e2031d06283117d36e9b564d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borrower details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Borrowers"
                ],
                "summary": "Update borrower",
                "description": "Update borrower profile",
                "operationId": "a06f4ab2bafc77e7e90343231ea424f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "middle_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "suffix": {
                                        "type": "string"
                                    },
                                    "birthdate": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "civil_status": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "contact_number": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "employer_or_business": {
                                        "type": "string"
                                    },
                                    "monthly_income": {
                                        "type": "number"
                                    },
                                    "branch_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Borrower updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Borrowers"
                ],
                "summary": "Delete borrower",
                "description": "Permanently delete a borrower and all related records",
                "operationId": "2896f3ba2578015d152a56a0108ec254",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borrower deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/borrowers/{id}/deactivate": {
            "patch": {
                "tags": [
                    "Borrowers"
                ],
                "summary": "Deactivate borrower",
                "operationId": "c7213e89c2a5363f42596347f6563d79",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borrower deactivated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/borrowers/{id}/reactivate": {
            "patch": {
                "tags": [
                    "Borrowers"
                ],
                "summary": "Reactivate borrower",
                "operationId": "3a003151d236f6cddbc42533095f685b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borrower reactivated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/borrowers/{id}/photo": {
            "post": {
                "tags": [
                    "Borrowers"
                ],
                "summary": "Upload borrower photo",
                "operationId": "1fdc6035343d82713aa8cdd9e6d271a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "photo"
                                ],
                                "properties": {
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo uploaded"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Borrowers"
                ],
                "summary": "Delete borrower photo",
                "operationId": "57b1eac6833dc6e29948060aeab0d01b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Photo deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}